// server/src/types/order.type.ts
import { BasicProduct, Product } from './product.type';
import { BasicCustomer, Customer } from './customer.type';

export interface BasicOrder {
  product: BasicProduct;
  customer: BasicCustomer;
  productQuantity: number;
}

export interface Order extends BasicOrder {
  orderId: number;
}

export interface OrderWithDetails extends Order {
  product: Product;
  customer: Customer;
}